requirements.txt
openai>=1.0.0
Pillow>=10.0.0
py-xmind16>=0.1.0
